+++ client/src/styles/shared/EntityPageSection.ts
@@ -40,7 +40,7 @@ const EntityPageSection = styled.div<IProps>`
 
   &.results,
   &.resultsAlert {
-    box-shadow: 0px 4px 5px 0px ${theme.color.black20};
+    box-shadow: 0px 2px 5px 5px ${theme.color.black10};
   }
 
   &.resultsEntityPageSection {