+++ client/src/features/timeline/Graph.tsx
@@ -11,6 +11,7 @@ import {
   ResponsiveContainer,
   Brush,
 } from 'recharts'
+import { isNull } from 'lodash'
 
 import theme from '../../styles/theme'
 import {
@@ -20,8 +21,10 @@ import {
 import { IHalLinks } from '../../types/IHalLinks'
 import TimelineParser from '../../lib/parse/timeline/TimelineParser'
 import useResizeableWindow from '../../lib/hooks/useResizeableWindow'
+import { facetNameMap } from '../../config/timeline'
 
 import CustomTooltip from './CustomTooltip'
+import CustomLegend from './CustomLegend'
 
 interface IProps {
   timelineData: ITimelinesTransformed
@@ -68,6 +71,20 @@ const Graph: React.FC<IProps> = ({
     window.innerWidth < theme.breakpoints.md,
   )
   const [graphData, setGraphData] = useState<Array<IGraphTimelineData>>([])
+  const [hoveredLegend, setHoveredLegend] = useState<string | null>(null)
+  const [selectedLegend, setSelectedLegend] = useState<string | null>(null)
+  // Get the relationships that are used in the timeline data to determine which relationships to show in the legend
+  const relationshipsToShowInLegend =
+    TimelineParser.getFacetsUsedForLegend(timelineData)
+  const activeLegend = hoveredLegend || selectedLegend
+
+  const handleOnHover = (value: string | null): void => {
+    setHoveredLegend(value)
+  }
+
+  const handleOnClick = (value: string | null): void => {
+    setSelectedLegend((currentValue) => (currentValue === value ? null : value))
+  }
 
   useEffect(() => {
     setGraphData(
@@ -84,22 +101,7 @@ const Graph: React.FC<IProps> = ({
     )
   }, [yearsArray, timelineData])
 
-  const facetNameMap: Map<string, string> = new Map([
-    ['itemProductionDate', 'Objects Produced'],
-    ['itemEncounteredDate', 'Objects Encountered'],
-    ['workCreationDate', 'Works Created'],
-    ['workPublicationDate', 'Works Published'],
-    ['workCreationOrPublicationDate', 'Works About'],
-    ['setAboutDate', 'Collections About'],
-    ['setCreationDate', 'Collections Created'],
-    ['setPublicationDate', 'Collections Published'],
-  ])
-
   useResizeableWindow(setIsMobile)
-
-  const renderLegendText = (value: string): any => (
-    <span style={{ color: 'black', fontWeight: '300' }}>{value}</span>
-  )
 
   const getShape = (props: any): ReactElement<any, any> => {
     const { fill, x, y, width, height } = props
@@ -147,96 +149,56 @@ const Graph: React.FC<IProps> = ({
           />
           <Legend
             layout={isMobile ? 'vertical' : 'horizontal'}
-            formatter={renderLegendText}
-          />
-          <Bar
-            dataKey="itemProductionDate.totalItems"
-            stackId="a"
-            fill={theme.color.graphs.produced}
-            name={
-              facetNameMap.get('itemProductionDate') || 'itemProductionDate'
+            content={
+              <CustomLegend
+                payload={undefined}
+                activeLegend={activeLegend}
+                selectedLegend={selectedLegend}
+                handleOnHover={handleOnHover}
+                handleOnClick={handleOnClick}
+              />
             }
-            yAxisId="total"
-            shape={(p: any) => getShape(p)}
           />
-          <Bar
-            dataKey="itemEncounteredDate.totalItems"
-            stackId="a"
-            fill={theme.color.graphs.encounter}
-            name={
-              facetNameMap.get('itemEncounteredDate') || 'itemEncounteredDate'
+          {Array.from(facetNameMap.entries()).map(([facetKey, facetLabel]) => {
+            if (!relationshipsToShowInLegend.includes(facetKey)) {
+              return null
             }
-            yAxisId="total"
-            shape={(p: any) => getShape(p)}
-          />
-          <Bar
-            dataKey="workCreationDate.totalItems"
-            stackId="a"
-            fill={theme.color.graphs.created}
-            name={facetNameMap.get('workCreationDate') || 'workCreationDate'}
-            yAxisId="total"
-            shape={(p: any) => getShape(p)}
-          />
-          <Bar
-            dataKey="workPublicationDate.totalItems"
-            stackId="a"
-            fill={theme.color.graphs.published}
-            name={
-              facetNameMap.get('workPublicationDate') || 'workPublicationDate'
+
+            let defaultLegend = 'focused'
+            if (isNull(activeLegend)) {
+              defaultLegend = 'focused'
+            } else if (activeLegend === facetKey) {
+              defaultLegend = 'focused'
+            } else {
+              defaultLegend = 'unFocused'
             }
-            yAxisId="total"
-            shape={(p: any) => getShape(p)}
+            return (
+              <Bar
+                key={facetKey}
+                dataKey={`${facetKey}.totalItems`}
+                stackId="a"
+                fill={
+                  theme.color.graphs[
+                    facetKey as keyof typeof theme.color.graphs
+                  ][defaultLegend as 'focused' | 'unFocused']
+                }
+                name={facetLabel || facetKey}
+                yAxisId="total"
+                shape={(p: any) => getShape(p)}
+              />
+            )
+          })}
+          <Brush
+            dataKey="year"
+            stroke={theme.color.primary.blue}
+            startIndex={startIndex}
+            endIndex={endIndex}
+            onDragEnd={(e: any) => {
+              const start = e.startIndex
+              const end = e.endIndex
+              handleRangeChange(start, end)
+            }}
           />
-          <Bar
-            dataKey="workCreationOrPublicationDate.totalItems"
-            stackId="a"
-            fill={theme.color.graphs.about}
-            name={
-              facetNameMap.get('workCreationOrPublicationDate') ||
-              'workCreationOrPublicationDate'
-            }
-            yAxisId="total"
-            shape={(p: any) => getShape(p)}
-          />
-          <Bar
-            dataKey="setAboutDate.totalItems"
-            stackId="a"
-            fill={theme.color.graphs.setAbout}
-            name={facetNameMap.get('setAboutDate') || 'setAboutDate'}
-            yAxisId="total"
-            shape={(p: any) => getShape(p)}
-          />
-          <Bar
-            dataKey="setCreationDate.totalItems"
-            stackId="a"
-            fill={theme.color.graphs.setCreated}
-            name={facetNameMap.get('setCreationDate') || 'setCreationDate'}
-            yAxisId="total"
-            shape={(p: any) => getShape(p)}
-          />
-          <Bar
-            dataKey="setPublicationDate.totalItems"
-            stackId="a"
-            fill={theme.color.graphs.setPublished}
-            name={
-              facetNameMap.get('setPublicationDate') || 'setPublicationDate'
-            }
-            yAxisId="total"
-            shape={(p: any) => getShape(p)}
-          />
-          {yearsArray.length > 1 && (
-            <Brush
-              dataKey="year"
-              stroke={theme.color.primary.blue}
-              startIndex={startIndex}
-              endIndex={endIndex}
-              onDragEnd={(e: any) => {
-                const start = e.startIndex
-                const end = e.endIndex
-                handleRangeChange(start, end)
-              }}
-            />
-          )}
         </BarChart>
       </ResponsiveContainer>
     </div>